requirements.txt
fastapi==0.110.0
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.9
starlette==0.36.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.4.0+cpu
uvicorn==0.28.0
